bff/src/test/java/bg/com/bo/bff/provider/TransferOtherAccountProviderTest.java
package bg.com.bo.bff.provider;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import static org.springframework.test.util.ReflectionTestUtils.setField;

import com.fasterxml.jackson.databind.ObjectMapper;

import org.apache.http.HttpEntity;
import org.apache.http.StatusLine;
import org.apache.http.client.methods.CloseableHttpResponse;
import org.apache.http.client.methods.HttpPost;
import org.apache.http.impl.client.CloseableHttpClient;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.Spy;
import org.mockito.junit.jupiter.MockitoExtension;

import java.io.ByteArrayInputStream;
import java.io.IOException;
import java.io.InputStream;
import java.util.LinkedHashSet;
import java.util.Set;

import bg.com.bo.bff.config.MiddlewareConfig;
import bg.com.bo.bff.config.exception.GenericException;
import bg.com.bo.bff.controllers.models.TransferRequestFixture;
import bg.com.bo.bff.controllers.models.TransferResponseFixture;
import bg.com.bo.bff.controllers.response.TransferResponse;
import bg.com.bo.bff.model.ClientToken;
import bg.com.bo.bff.model.interfaces.IHttpClientFactory;
import bg.com.bo.bff.provider.interfaces.ITokenMiddlewareProvider;
import bg.com.bo.bff.provider.models.MiddlewareConfigFixture;
import bg.com.bo.bff.provider.response.ApiDataResponse;
import bg.com.bo.bff.provider.response.ApiErrorResponse;
import bg.com.bo.bff.provider.response.ErrorDetailResponse;

@ExtendWith(MockitoExtension.class)
class TransferOtherAccountProviderTest {


    @Spy
    @InjectMocks
    TransferOtherAccountProvider provider;

    @Mock
    ITokenMiddlewareProvider tokenMiddlewareProvider;

    @Mock
    MiddlewareConfig middlewareConfig;

    @Mock
    IHttpClientFactory httpClientFactory;

    @Mock
    CloseableHttpClient closeableHttpClientMock;

    @Mock
    CloseableHttpResponse closeableHttpGetResponseMock;

    @Mock
    HttpEntity httpEntityMock;

    @Mock
    StatusLine statusLineMock;


    @BeforeEach
    void setUp() {
        provider = new TransferOtherAccountProvider(httpClientFactory, middlewareConfig, tokenMiddlewareProvider);
        setField(provider, "middlewareConfig", MiddlewareConfigFixture.withDefault(), MiddlewareConfig.class);
    }

    @Test
    void transfer() throws IOException {
        ClientToken clientToken = new ClientToken();
        clientToken.setAccessToken("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1bmlxdWVfbmFtZSI6IjdjNjlmZjNjZjdlNjE5MWU2NGUxMmZhMGVlNmM2ZWNiNTBiODkyY2E5NzIyMmJmZmMxMTc0Yzg5ZTcwNGM5NDQiLCJyb2xlIjoiMTY5YjRlM2IyNzhiYzAzYzZjNWUzNTQ4MDk5ZDUyZTk1MzRmZDRkNjhmMTM0MmEzNzM0OWFjYjQ1NWQ2ZWRjOCIsImdyb3Vwc2lkIjoiMmZhN2MxYjljNjE1ZmU5NThjYmFkODAyNDQzMGNjYjM3ZGE5YTEyMGExMjJiYWI0ZDEyMTFjMGQ3MDMyMTEwYiIsInByaW1hcnlzaWQiOiIyNGI4YjIxNTE1ZTU4ZDdkYTJiZTE1ZWFkZjBhODUyODg5NjEyNTMzODI4ZjkxNDA2YWJmNjRmYjgyYTViNjE2IiwibmJmIjoxNjk5OTIyMzg5LCJleHAiOjE2OTk5MjQxODksImlhdCI6MTY5OTkyMjM4OSwiaXNzIjoiaHR0cDovL3NlcnZpY2lvcy5iZ2EuY29tLmJvIiwiYXVkIjoiaHR0cDovL3NlcnZpY2lvcy5iZ2EuY29tLmJvIn0.J-Is_mRLEwwn8Z-RyAe40t0TpkLoppTE7roWe0zXFoc");
        clientToken.setExpiresIn(1699924189);
        when(tokenMiddlewareProvider.generateAccountAccessToken(any())).thenReturn(clientToken);
        Mockito.when(httpClientFactory.create()).thenReturn(closeableHttpClientMock);
        Mockito.when(closeableHttpClientMock.execute(Mockito.any(HttpPost.class))).thenReturn(closeableHttpGetResponseMock);
        Mockito.when(closeableHttpGetResponseMock.getEntity()).thenReturn(httpEntityMock);
        Mockito.when(closeableHttpGetResponseMock.getStatusLine()).thenReturn(statusLineMock);
        Mockito.when(statusLineMock.getStatusCode()).thenReturn(200);

        ObjectMapper objectMapper = new ObjectMapper();

        String json = objectMapper.writeValueAsString(ApiDataResponse.builder().data(TransferResponseFixture.withDefault()).build());
        InputStream inputStream = new ByteArrayInputStream(json.getBytes());

        Mockito.when(httpEntityMock.getContent()).thenReturn(inputStream);


        TransferResponse result = provider.transfer("123455", TransferRequestFixture.withDefault());

        assertNotNull(result);
        verify(tokenMiddlewareProvider).generateAccountAccessToken(any());

    }

    @Test
    void transferError() throws IOException {
        ClientToken clientToken = new ClientToken();
        clientToken.setAccessToken("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1bmlxdWVfbmFtZSI6IjdjNjlmZjNjZjdlNjE5MWU2NGUxMmZhMGVlNmM2ZWNiNTBiODkyY2E5NzIyMmJmZmMxMTc0Yzg5ZTcwNGM5NDQiLCJyb2xlIjoiMTY5YjRlM2IyNzhiYzAzYzZjNWUzNTQ4MDk5ZDUyZTk1MzRmZDRkNjhmMTM0MmEzNzM0OWFjYjQ1NWQ2ZWRjOCIsImdyb3Vwc2lkIjoiMmZhN2MxYjljNjE1ZmU5NThjYmFkODAyNDQzMGNjYjM3ZGE5YTEyMGExMjJiYWI0ZDEyMTFjMGQ3MDMyMTEwYiIsInByaW1hcnlzaWQiOiIyNGI4YjIxNTE1ZTU4ZDdkYTJiZTE1ZWFkZjBhODUyODg5NjEyNTMzODI4ZjkxNDA2YWJmNjRmYjgyYTViNjE2IiwibmJmIjoxNjk5OTIyMzg5LCJleHAiOjE2OTk5MjQxODksImlhdCI6MTY5OTkyMjM4OSwiaXNzIjoiaHR0cDovL3NlcnZpY2lvcy5iZ2EuY29tLmJvIiwiYXVkIjoiaHR0cDovL3NlcnZpY2lvcy5iZ2EuY29tLmJvIn0.J-Is_mRLEwwn8Z-RyAe40t0TpkLoppTE7roWe0zXFoc");
        clientToken.setExpiresIn(1699924189);
        when(tokenMiddlewareProvider.generateAccountAccessToken(any())).thenReturn(clientToken);
        Mockito.when(httpClientFactory.create()).thenReturn(closeableHttpClientMock);
        Mockito.when(closeableHttpClientMock.execute(Mockito.any(HttpPost.class))).thenReturn(closeableHttpGetResponseMock);
        Mockito.when(closeableHttpGetResponseMock.getEntity()).thenReturn(httpEntityMock);
        Mockito.when(closeableHttpGetResponseMock.getStatusLine()).thenReturn(statusLineMock);
        Mockito.when(statusLineMock.getStatusCode()).thenReturn(406);

        Set<ErrorDetailResponse> responseList = new LinkedHashSet<>();
        responseList.add(ErrorDetailResponse.builder().code("MDW").description("Error").build());
        ApiErrorResponse errorResponse = ApiErrorResponse.builder()
                .errorType("Technical")
                .code(406)
                .errorDetail(responseList).build();

        ObjectMapper objectMapper = new ObjectMapper();
        String json = objectMapper.writeValueAsString(errorResponse);
        InputStream inputStream = new ByteArrayInputStream(json.getBytes());

        Mockito.when(httpEntityMock.getContent()).thenReturn(inputStream);


        GenericException response = assertThrows(GenericException.class, () -> provider.transfer("123455", TransferRequestFixture.withDefault()));


        assertNotNull(response);

        assertEquals(response.getStatus().value(), errorResponse.getCode());

        verify(tokenMiddlewareProvider).generateAccountAccessToken(any());

    }
}